requirements.txt
python-telegram-bot[job-queue]>=21.3
sqlalchemy>=2.0.25
python-dotenv>=1.0.0
aiohttp>=3.9.0
